Move the application model into a separate directory
