+++ generator/src/cmd/generate.go
@@ -13,11 +13,13 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+
 package cmd
 
 import (
-	"application-generator/src/pkg/generate"
-	"application-generator/src/pkg/model"
+	"cloud-native-app-simulator/generator/src/pkg/generate"
+	"cloud-native-app-simulator/model"
+
 	"bufio"
 	"fmt"
 	"os"
