+++ generator/src/pkg/generate/generate.go
@@ -13,11 +13,13 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+
 package generate
 
 import (
-	"application-generator/src/pkg/model"
-	s "application-generator/src/pkg/service"
+	s "cloud-native-app-simulator/generator/src/pkg/service"
+	"cloud-native-app-simulator/model"
+
 	"bytes"
 	"encoding/json"
 	"fmt"
