+++ generator/src/pkg/service/util.go
@@ -13,10 +13,12 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+
 package service
 
 import (
-	"application-generator/src/pkg/model"
+	"cloud-native-app-simulator/model"
+
 	"fmt"
 	"strconv"
 )
