+++ go.mod
@@ -1,4 +1,4 @@
-module application-generator
+module cloud-native-app-simulator
 
 go 1.20
 
@@ -13,5 +13,5 @@ require (
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
 	gopkg.in/inf.v0 v0.9.1 // indirect
-	k8s.io/utils v0.0.0-20230505201702-9f6742963106 // indirect
+	k8s.io/utils v0.0.0-20230209194617-a36077c30491 // indirect
 )
